Replace with a Go program
My doorbell circuit is low voltage AC, and the optocoupler will be
either off, or a pulse of events (because of AC).

Using Go makes it easier (for me!) to write the logic to debounce and
filter button presses.

Filtering is critical, because the GPIO pin is detecting rising and
falling edges when electrical appliances power on/off in the house!