Fix doorbell notify
Revert back to a simple "pause music" doorbell behaviour and put this
back into version-controlled code. I'd been experimenting with a more
clever system that played doorbell tones with a python script but it was
not reliable and I want something reliable and working to revert to in
version control.